requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.27
mysql-connector-python==8.0.27
pydantic==1.8.2  
passlib==1.7.4
python-jose==3.3.0
